requirements.txt
azure-cognitiveservices-speech
beautifulsoup4
